% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmer2.R
\name{lmer2}
\alias{lmer2}
\title{Wrapper for lme4's \code{lmer} Function}
\usage{
lmer2(formula, data, onlyCoefs = FALSE, start = NULL,
  adaptiveStart = TRUE, lmerControls = NULL, warningsToNA = FALSE,
  errorsToNA = FALSE, triggerConvWarning = FALSE, ...)
}
\arguments{
\item{formula}{same as \code{\link[lme4]{lmer}} \code{formula}.}

\item{data}{same as \code{\link[lme4]{lmer}} \code{data}.}

\item{onlyCoefs}{logical: whether to only return model coefficients.}

\item{start}{same as \code{\link[lme4]{lmer}} \code{start}.}

\item{adaptiveStart}{logical: whether to use ending theta values of the
previous iteration as \code{start} values for the current iteration.}

\item{lmerControls}{\code{\link[base]{list}} of
\code{\link[lme4]{lmerControl}} functions to try when fitting the model.}

\item{warningsToNA}{logical: whether to return a matrix of NAs instead of
coefficients if a warning is raised when fitting the model.}

\item{errorsToNA}{logical: whether to return a matrix of NAs (with the same
shape as a coefficient matrix) if an error is raised when fitting the model.
The error is converted to a warning instead of unwinding the stack.}

\item{triggerConvWarning}{logical: raises a convergence warning.}

\item{...}{other potential arguments passed to \code{\link[lme4]{lmer}}.}
}
\value{
If \code{onlyCoef==TRUE} then the output will be a matrix of betas,
standard errors, and t-values. Otherwise, the ouput will be a list that
contains this matrix as well as the standard ouput of
\code{\link[lme4]{lmer}} (an object of class \code{lmerMod}) and a list of
any messages, warnings, and errors generated when fitting the model.
}
\description{
Fit linear mixed-effects models using \code{\link[lme4]{lmer}} with access
to additional features that support automated refitting and condition
handling.
}
\examples{
library(lme4)
data(sleepstudy)

## try these optimizers in order
controls <- list(lmerControl(optimizer='nloptwrap',
                             optCtrl=list(maxfun=1e9,
                             algorithm="NLOPT_LN_BOBYQA")),

                 lmerControl(optimizer="bobyqa",
                             optCtrl=list(maxfun=1e9)),

                 lmerControl(optimizer="Nelder_Mead",
                             optCtrl=list(maxfun=1e9)))

## fit the model
coefs <- lmer2(Reaction ~ Days + (Days | Subject), sleepstudy, TRUE,
               lmerControls=controls)
}

